add option to show images for like/nope view
Arbitrary images are now supported in the place of like/nope text. This
also fixes a few warnings found in Xcode 6.3 regarding fabsf/fabs.

+++ MDCSwipeToChoose/Public/Views/UIView+MDCSwipeToChoose.m
@@ -171,7 +171,7 @@ const void * const MDCViewStateKey = &MDCViewStateKey;
 
 - (void)mdc_executeOnPanBlockForTranslation:(CGPoint)translation {
     if (self.mdc_options.onPan) {
-        CGFloat thresholdRatio = MIN(1.f, fabsf(translation.x)/self.mdc_options.threshold);
+        CGFloat thresholdRatio = MIN(1.f, fabs(translation.x)/self.mdc_options.threshold);
 
         MDCSwipeDirection direction = MDCSwipeDirectionNone;
         if (translation.x > 0.f) {
